Rename group_vars to host_vars
The variables are sourced manually anyway instead of relying on
Ansible's default mechanism, since the host is always `localhost` in our
case, so we pick var file based on system facts.

But host_vars describes better of what is actually stored in var files
(variables specific to host where Arch Linux is installed) so it still
makes more sense.